+++ src/hooks/connection.ts
@@ -23,6 +23,8 @@ import { fuzzyMatch } from "~/util/helpers";
 import { useConnectionAndView } from "./routing";
 import { useStable } from "./stable";
 
+export type ConnectionType = "cloud" | "local" | "sandbox";
+
 /**
  * Returns whether Surrealist is connected to a database
  */
@@ -76,6 +78,28 @@ export function useConnections<T>(selector: (con: Connection) => T): T[] {
 			return s.connections.map(selector);
 		}),
 	);
+}
+
+/**
+ * Returns the type of the current connection
+ */
+export function useConnectionType() {
+	const [connectionId, instanceId] = useConnection((c) => [
+		c?.id ?? "",
+		c?.authentication.cloudInstance ?? "",
+	]);
+
+	switch (true) {
+		case !!instanceId: {
+			return "cloud";
+		}
+		case connectionId === SANDBOX: {
+			return "sandbox";
+		}
+		default: {
+			return "local";
+		}
+	}
 }
 
 /**
