+++ src/screens/surrealist/components/ToolbarBreadcrumbs/index.tsx
@@ -1,22 +1,12 @@
 import { Breadcrumbs, Text } from "@mantine/core";
+import { Fragment } from "react";
 import { BreadcrumbCrumb } from "~/components/BreadcrumbCrumb";
-import { SANDBOX } from "~/constants";
-import { useIsConnected } from "~/hooks/connection";
-import { useConnectionAndView } from "~/hooks/routing";
 import { useInterfaceStore } from "~/stores/interface";
-import { ConnectionCrumb } from "./connection";
-import { DatabaseCrumb } from "./database";
 
 export function ToolbarBreadcrumbs() {
 	const pageBreadcrumbs = useInterfaceStore((s) => s.pageBreadcrumbs);
 
-	const [connection] = useConnectionAndView();
-	const isConnected = useIsConnected();
-
-	const isSandbox = connection === SANDBOX;
-	const showDatabase = !isSandbox && connection && isConnected;
-
-	if (pageBreadcrumbs.length === 0 && !connection) {
+	if (pageBreadcrumbs.length === 0) {
 		return null;
 	}
 
@@ -33,15 +23,10 @@ export function ToolbarBreadcrumbs() {
 			}
 		>
 			{pageBreadcrumbs.map((item, index) => (
-				<BreadcrumbCrumb
-					key={`${item.label}-${index}`}
-					item={item}
-				/>
+				<Fragment key={`${item.label}-${index}`}>
+					{item.content ?? <BreadcrumbCrumb item={item} />}
+				</Fragment>
 			))}
-
-			{connection && <ConnectionCrumb />}
-
-			{showDatabase && <DatabaseCrumb />}
 		</Breadcrumbs>
 	);
 }
